Conflicts:
	src/backend/nodes/outfuncs.c
	src/backend/optimizer/plan/createplan.c
	src/include/optimizer/cost.h

+++ src/include/executor/spi.h
@@ -21,8 +21,8 @@
 typedef struct SPITupleTable
 {
 	MemoryContext tuptabcxt;	/* memory context of result table */
-	uint32		alloced;		/* # of alloced vals */
-	uint32		free;			/* # of free vals */
+	uint64		alloced;		/* # of alloced vals */
+	uint64		free;			/* # of free vals */
 	TupleDesc	tupdesc;		/* tuple descriptor */
 	HeapTuple  *vals;			/* tuples */
 	slist_node	next;			/* link for internal bookkeeping */
@@ -63,7 +63,7 @@ typedef struct _SPI_plan *SPIPlanPtr;
 #define SPI_OK_TSR_REGISTER		15
 #define SPI_OK_TSR_UNREGISTER	16
 
-extern PGDLLIMPORT uint32 SPI_processed;
+extern PGDLLIMPORT uint64 SPI_processed;
 extern PGDLLIMPORT Oid SPI_lastoid;
 extern PGDLLIMPORT SPITupleTable *SPI_tuptable;
 extern PGDLLIMPORT int SPI_result;
